+++ src/main/kotlin/de/bund/digitalservice/useid/wellknown/WellKnownController.kt
@@ -1,4 +1,4 @@
-package de.bund.digitalservice.useid.statics
+package de.bund.digitalservice.useid.wellknown
 
 import mu.KotlinLogging
 import org.springframework.http.MediaType
